+++ austin-common/src/main/java/com/java3y/austin/enums/MessageType.java
@@ -30,7 +30,7 @@ public enum MessageType {
     /**
      * 英文标识
      */
-    private String code_en;
+    private String codeEn;
 
 
     /**
