+++ austin-handler/src/main/java/com/java3y/austin/handler/EmailHandler.java
@@ -21,7 +21,7 @@ import org.springframework.stereotype.Component;
  */
 @Component
 @Slf4j
-public class EmailHandler extends Handler {
+public class EmailHandler extends BaseHandler implements Handler {
 
     private static final String EMAIL_ACCOUNT_KEY = "emailAccount";
     private static final String PREFIX = "email_";
