+++ austin-handler/src/main/java/com/java3y/austin/handler/SmsHandler.java
@@ -24,7 +24,7 @@ import java.util.List;
  */
 @Component
 @Slf4j
-public class SmsHandler extends Handler {
+public class SmsHandler extends BaseHandler implements Handler  {
 
     public SmsHandler() {
         channelCode = ChannelType.SMS.getCode();
