+++ austin-handler/src/main/java/com/java3y/austin/service/deduplication/service/AbstractDeduplicationService.java
@@ -94,7 +94,7 @@ public abstract class AbstractDeduplicationService implements DeduplicationServi
             }
         }
         if (CollUtil.isNotEmpty(keyValues)) {
-            redisUtils.pipelineSetEX(keyValues, param.getDeduplicationTime());
+            redisUtils.pipelineSetEx(keyValues, param.getDeduplicationTime());
         }
     }
