+++ austin-support/src/main/java/com/java3y/austin/utils/LogUtils.java
@@ -8,10 +8,9 @@ import com.java3y.austin.domain.LogParam;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
 
-import java.util.Date;
-
 /**
  * 所有的日志都存在
+ * @author 3y
  */
 @Slf4j
 @Component
