+++ austin-support/src/main/java/com/java3y/austin/utils/TaskInfoUtils.java
@@ -13,6 +13,7 @@ import java.util.Date;
 public class TaskInfoUtils {
 
     private static final int TYPE_FLAG = 1000000;
+    private static final char PARAM = '?';
 
     /**
      * 生成BusinessId
@@ -30,7 +31,7 @@ public class TaskInfoUtils {
     public static String generateUrl(String url, Long templateId, Integer templateType) {
         url = url.trim();
         Long businessId = generateBusinessId(templateId, templateType);
-        if (url.indexOf('?') == -1) {
+        if (url.indexOf(PARAM) == -1) {
             return url + "?track_code_bid=" + businessId;
         } else {
             return url + "&track_code_bid=" + businessId;
